feat(components): [virtual-table] columns
- Define column props and types
- Rearranged types
