+++ packages/components/table-v2/src/constants.ts
@@ -1,0 +1,1 @@
+export const sortOrders = ['asc', 'desc'] as const
